Remove unecessary fields in Species class
